% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defi2.R
\docType{data}
\name{defi2}
\alias{defi2}
\title{Eurostat data}
\format{a \code{\link[tibble]{tibble}} with 12 columns and 1190 rows.}
\source{
Eurostat, see \href{??}{here}.
}
\description{
This dataset is made of ...
}
\note{
This dataset is distributed by...
}
\keyword{datasets}
