% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_defi2.R
\name{make_defi2}
\alias{make_defi2}
\title{Creation of the 'defi2' dataset}
\usage{
make_defi2(path = ".", save_it = FALSE)
}
\arguments{
\item{path}{character. Data once created are saved in the folder 
\code{file.path(path, "data")}.}

\item{save_it}{logical. If \code{TRUE}, the result is saved as an \code{.RData} file in the 
folder \code{file.path(path, "data")}.}
}
\value{
Returns invisibly the tibble created.
}
\description{
The function \code{make_defi2} creates the \code{\link[sdo]{defi2}} 
dataset from the file \code{Defi2_IndicCle_Eurostat.xlsx} downloaded 
\href{https://}{here}.
}
\examples{
\dontrun{
dir.create("data-raw")
make_defi2(save_it = TRUE)
}

}
\seealso{
\code{\link[sdo]{defi2}}
}
